User 엔티티 수정, 대시보드 백에 필요한 데이터 추가 작업중

+++ src/main/java/com/elearning/instructor/dto/dashboard/ProgressStatusDto.java
@@ -14,6 +14,6 @@ import lombok.NoArgsConstructor;
 @Builder
 public class ProgressStatusDto {
   private String rangeLabel;  // 예: "0~25%"
-  private int studentCount;   // 해당 구간에 속하는 수강생 수
+  private Long studentCount;   // 해당 구간에 속하는 수강생 수
 }
